JSON: Fix Infinite loop bug in Serialize
This patch updates the JSON module in the framework to fix a bug in `Serialize`
JSON::Container and JSON::ArrayType.

Serialize method was getting stuck in infinite loop in some scenarios.

On Analyzing it was found that the issue was occuring when Serialize method
fails to add `}` or `]` becuase the buffer is full.

The logic will set the offset to ~0 when it wants to add `}` or `]`.
And if we cant add either of those characters and we are in a nested level
of parsing we will add PARSE to ~0 which will overflow ~0 and the offset will
become PARSE - 1. After which the parsing logic will be in an invalid state.

As a fix we are ensuring that we don't overflow ~0. Also we are using END_MARKER
to indicate that only `}` or `]` is pending.

And if we get END_MARKER for a new buffer we will replace it with ~0 if there is
no more data to parse.

Signed-off-by: harikris2012 <harikris2012@gmail.com>